Adding class queue count extraction interface and minor correction.
Added the interface to extract the class queue current number of frames from the attributes.
Added a more performant class queue current number of frames.
Added a way to extract the i_cnt from the CEETM congestion group.
Added a more performant way to query the i_cnt from the CEETM congestion group.
Fix for the extraction of the statistic for class 16 to 31.

Signed-off-by: Youri Querry <youri.querry_1@nxp.com>