examples/2fwd-crypto-event: fix keys data
The cipher keys and auth keys given in command line
are not being used for actual PDCP processing. Application is
parsing and storing the user given keys but passing 0 values
for the encryption/decryption and authentication.

This patch fixes this issue by passing the user given keys
for PDCP processing.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
DPDK-3092